Fix a regression on beta Rust
This bit of code relies on behavior that will no longer work in Rust 1.79,
which is currently in beta. Since this crate was the only regression, the
behavior change was considered acceptable, but does mean that this crate has a
minor breakage in tests. 

Add a concrete type to the vector to correct for this behavior change. 

See the context here:
https://github.com/rust-lang/rust/issues/125198